[clangd] Support iwyu pragma: no_include
This patch is one of the steps trying to optimize experience
of header insertion in code completion.
IWYU supports `no_include` pragma that helps user to get rid
of specified headers in main file.
The patch consists of two parts: Code completion part that
prevent auto insertion if the header to be inserted is under
the guard of `no_include` pragma. The other part is for include
cleaner, which emits diagnostic message pointing out that user
is including `no_include` headers.

Differential Revision: https://reviews.llvm.org/D143319